pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mousetube_api"
version = "1.0"
description = "mousetube allows the exchange of mouse vocalisation recording files."
authors = [
    { name = "Nicolas Torquet", email = "nicolas.torquet@igbmc.com" },
    { name = "Laurent Bouri", email = "laurent.bouri@igbmc.fr" },
]
dependencies = [
    "Django==5.2.8",
    "djangorestframework==3.16.0",
    "djoser==2.3.1",
    "django-cors-headers==4.7.0",
    "mysqlclient==2.2.7",
    "django-environ==0.12.0",
    "drf-spectacular==0.28.0",
    "uvicorn==0.29.0",
    "gunicorn==23.0.0",
    "mkdocs-material>=8.0",
    "mkdocstrings[python]>=0.11",
    "django-countries>=7.5",
    "setuptools==78.1.1",
    "pillow==11.3.0",
    "celery[redis]==5.4.0",
    "redis==5.0.8",
    "django-celery-results==2.5.1",
    "celery-progress==0.5",
    "django-extensions==4.1",
    "django-allauth>=65.9.0",
    "django-filter>=25.1",
    "soundfile==0.13.1",
]
license = { text = "GPL-3.0" }

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[project.scripts]
mousetube_api = "mousetube_api:manage"

[tool.setuptools.packages.find]
include = ["mousetube_api*"]

[tool.dependencies]
mkdocs-material = ">=8.0"
mkdocstrings = { version = ">=0.11", extras = ["python"] }
